Adding in a prototype MafOutputRenderer for Funcotator (and many fixes) (#4472)
* Adding in a MafOutputRenderer for Funcotator.

Added in argument for MAF out.
Added more "required" MAF fields.
Added Funcotation::getDataSourceName
Added contig name overrides for hg19 VS B37
Added in code to convert VCF INDEL positions to MAF INDEL positions.
Added start/stop positions for IGRs.
Added argument to ignore filtered variants at the front of processing to save time.
Added a script to fully retrieve the COSMIC data sources.

Fixed how the MafOutputRenderer handles mapping fields to values.
Fixed a bug in LocatableXsv and COSMIC parsers (was missing name and
    version).
In Gencode: Now TumorSeqAllele1 is the refAllele, not the AltAllele.
Fixed some problems with VCF output.
Updated VCF outputs to have better header info.
Refactored header output for OutputRenderers.
Changed the logic for creating alt protein sequences.
Fixed a bug in the LocatableXsvFuncotationFactory that caused annotations to be incorrectly associated with a factory.
Fixed several bugs in the GencodeFuncotationFactory.
Fixed bugx in the handling of UTR variants.
Fixed the Transcript Selection Mode ordering.
Fixed an issue with splice sites.

Minor speed fix to GencodeFuncotationFactory.

Now CosmicFuncotationFactory opens the database in read-only mode.

Bugfix - now LocatableXsvFuncotationFactories use overrides.

Now the reference should properly align with ALL indels regardless of
length.

ReferenceContext now always rendered on + strand.

Now will create funcotations for transcripts without fasta sequences.

Minor changes to FuncotatorIntegrationTest.

- Added in more integration test files.  These are as yet unused - must
refactor the files themselves to actually reflect what should be correct
as far as produced funcotations.

- Updated LocatableXsvFuncotationFactoryUnitTest.java and SimpleKeyXsvFuncotationFactoryUnitTest.java
to reflect the change to funcotation factories to always produce the
expected funcotations (rather than only producing funcotations when
    there are data that match the target variant).

Fixed an issue with the new VariantClassification code.

Fixed issue #4410

Fixed #4022

Fixed #4420

Fixed #3922

* Removed tumorSeqAllele1 as an explicit field in GencodeFuncotation.

* Now GencodeFuncotation enums use maps for string<->enum lookups.

* Removed GencodeGtfFeature.GenomicStrand.  Replaced it with Strand from HTSJDK.

* Addressed more comments from reviewers.

* Now sorts the data sources before passing them to `OutputRenderer`s.

+++ scripts/funcotator/data_sources/cosmic/createSqliteCosmicDb.sh
@@ -5,10 +5,36 @@
 # This resulting database file will only contain records that have either
 # genome positions or protein positions so they can be matched to variants
 # by Funcotator.
-# To change the input file, change the `COSMIC_FILE` variable.
+# To change the input file, change the `COSMIC_FILE` variable or pass a path
+# in on the command-line.
+# To change the output file, change the `OUT_DB_FILE` variable or pass a path
+# in on the command-line for the COSMIC_FILE and a SECOND path for the OUT_DB_FILE
+
+################################################################################
+
+set -e
+
+################################################################################
 
 COSMIC_FILE=CosmicCompleteTargetedScreensMutantExport.tsv
 OUT_DB_FILE="Cosmic.db"
+
+################################################################################
+
+if [[ $# -gt 0 ]] ; then
+    COSMIC_FILE=$1
+fi
+
+if [[ $# -gt 1 ]] ; then
+    OUT_DB_FILE=$2
+fi
+
+if [ ! -f ${COSMIC_FILE} ] ; then
+    echo "ERROR: Given COSMIC file does not exist: ${COSMIC_FILE}" 1>&2
+    exit 1
+fi
+
+echo "Creating COSMIC database from file: ${COSMIC_FILE} -> ${OUT_DB_FILE}"
 
 [ -f ${OUT_DB_FILE} ] && rm ${OUT_DB_FILE}
 
@@ -22,3 +48,4 @@ CREATE INDEX GeneIndex ON Cosmic("Gene name");
 VACUUM;
 EOF
 
+echo 'DONE!'
